% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\docType{data}
\name{combo}
\alias{combo}
\title{BMI data}
\format{
A list containing 17 elements
}
\usage{
data(combo)
}
\description{
The COMBO study is a cross-sectional study designed to examine the association between micronutrients and gut microbiome composition, where the fecal samples of 98 healthy individuals from the University of Pennsylvania were collected along with demographic data like Body Mass Index (BMI), age, and sex.
The data are available at the Qiita repository under the accession ID: 101135.
}
\references{
Wu GD, Chen J, Hoffmann C, Bittinger K, Chen YY, Keilbaugh SA, Bewtra M,
Knights D, Walters WA, Knight R, Sinha R, Gilroy E, Gupta K, Baldassano R,
Nessel L, Li H, Bushman FD, Lewis JD. Linking long-term dietary patterns with
gut microbial enterotypes. \emph{Science}. 2011 Oct 7;334(6052):105-8.
doi: 10.1126/science.1208344. Epub 2011 Sep 1. PMID: 21885731; PMCID: PMC3368382.
}
\keyword{datasets}
