% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.genus.R
\docType{data}
\name{count.genus}
\alias{count.genus}
\title{CRC data (genus level)}
\format{
An integer matrix with 574 rows and 133 columns.
}
\source{
\href{https://github.com/zellerlab/crc_meta}{Zeller Lab}
}
\usage{
data(count.genus)
}
\description{
This dataset contains data from a meta-analysis of five geographically and technically diverse fecal shotgun metagenomic studies of colorectal cancer (CRC, n = 574). All raw sequencing data across studies were reprocessed using the same bioinformatics pipeline for taxonomic profiling.
The 5 studies involve subjects from 5 different countries. The sample size of the studies is 109, 127, 120, 114, and 104; and the number of cases and controls are roughly balanced in each study. This data-set contains p = 133 distinct genera found among all the investigated metagenomic CRC studies.
}
\keyword{datasets}
