% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_heatmap.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw Heatmap for Microbial Sample Taxa Matrix}
\usage{
draw_heatmap(X, title = "")
}
\arguments{
\item{X}{A numeric matrix representing microbial sample taxa data, where rows
represent samples and columns represent taxa.}

\item{title}{An optional title for the heatmap}
}
\value{
A ggplot object representing the heatmap. This can be further customized
or directly plotted.
}
\description{
This function creates a heatmap for a given microbial sample taxa matrix.
It is specifically designed for visualizing abundance patterns across
different samples and taxa in microbiome studies. The function applies
an arcsinh transformation to the data for normalization and better visualization
of abundance patterns, especially useful in handling highly skewed microbiome data.
}
\details{
The heatmap is generated using \code{ggplot2} and \code{reshape2} packages, with taxa
on the x-axis and samples on the y-axis. The color intensity in the heatmap
represents the arcsinh-transformed abundance of each taxa in each sample.
}
\examples{
# Create a sample matrix representing microbial sample taxa data
mat <- matrix(runif(20), nrow = 5)

# Draw heatmap
draw_heatmap(mat)

}
