% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKnockoff.R
\name{generateKnockoff}
\alias{generateKnockoff}
\title{Generate Knockoff Copy of Microbial Sample Data}
\usage{
generateKnockoff(X, Theta, Beta, seed = NULL)
}
\arguments{
\item{X}{A numeric matrix representing original microbial sample data.}

\item{Theta}{A numeric matrix representing cluster mixing probabilities.}

\item{Beta}{A numeric matrix representing feature proportions for each cluster.}

\item{seed}{An optional integer seed for reproducibility of random generation.}
}
\value{
A numeric matrix representing the knockoff copy of the microbial sample data.
}
\description{
This function generates a knockoff copy of microbial sample data given a matrix \code{X}
and matrices \code{Theta} and \code{Beta}. The function adjusts the column structure of \code{Beta}
to match \code{X}, generates samples based on \code{Theta} and \code{Beta}, and then compiles these
into a knockoff count matrix.
}
\examples{
X <- matrix(runif(40), nrow = 10)
colnames(X) <- paste("Taxa", 1:ncol(X))
Theta <- matrix(runif(30), nrow = 10)
Beta <- matrix(runif(20), nrow = 5)
knockoff_data <- generateKnockoff(X, Theta, Beta)

}
