% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormalize.R
\name{log_normalize}
\alias{log_normalize}
\title{Log-Normalization for Microbiome Compositional Data}
\usage{
log_normalize(X)
}
\arguments{
\item{X}{A numeric matrix to log-normalize.}
}
\value{
A matrix of the same dimensions as \code{X}, with each element
being the log-normalized value of the corresponding element in \code{X}.
The base of the logarithm used for normalization is e (natural logarithm).
}
\description{
This function performs log-normalization on a given matrix, typically used
in microbiome data analysis. In microbiome studies, data are often compositional
and contain many zero counts. Log-normalization, with the addition of a pseudo-count,
is a standard approach to handle zeros and maintain the compositional nature of the data.
This function takes a numeric matrix, adds a pseudo-count to zero values,
and then applies a log transformation, preserving the relative proportions
in the data.
}
\details{
The function first adds a pseudo-count of 0.5 to zero entries to handle zeros
in the data. It then divides each count by the total counts in its row (sample)
to make the data compositional. Finally, it applies a natural logarithm
transformation to the normalized data.
#'
}
\examples{
# Create a sample matrix
mat <- matrix(1:4, nrow = 2)

# Perform log-normalization
log_normalized_mat <- log_normalize(mat)

}
