% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinck.filter.R
\name{zinck.filter}
\alias{zinck.filter}
\title{Zinck Filter for Variable Selection}
\usage{
zinck.filter(
  X,
  X_tilde,
  Y,
  model,
  fdr = 0.1,
  offset = 1,
  seed = NULL,
  ntrees = 1000,
  tune_mtry = FALSE,
  mtry = NULL,
  metric = NULL,
  rftuning = FALSE,
  conservative = FALSE
)
}
\arguments{
\item{X}{An OTU matrix with dimensions \eqn{D \times p}.}

\item{X_tilde}{A knockoff matrix corresponding to X with the same dimensions.}

\item{Y}{The response variable, either continuous or binary.}

\item{model}{A string; the model to be used ("glmnet" or "Random Forest").}

\item{fdr}{Numeric; the false discovery rate. Default is 0.1}

\item{offset}{Numeric; either 0 or 1. Default is 1}

\item{seed}{Numeric; the seed for reproducibility.}

\item{ntrees}{Numeric; the number of trees for Random Forest. Default is 1000.}

\item{tune_mtry}{Logical; whether to tune mtry in Random Forest. Default is FALSE.}

\item{mtry}{Numeric; the number of variables randomly sampled as candidates at each split in Random Forest. Default is NULL (auto-tuned).}

\item{metric}{A string; the importance metric for the Random Forest, can be ("Accuracy" or "Gini"). Default is "Accuracy".}

\item{rftuning}{Logical; whether to fit a fast or a tuned implementation of a Random Forest. Default is FALSE.}

\item{conservative}{Logical; whether to use a conservative knockoff threshold or not. Default is FALSE.}
}
\value{
A vector of selected variables at a target false discovery rate
}
\description{
Performs variable selection by fitting the augmented set of features to the response, using a glmnet or Random Forest model
}
