% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_k.R
\name{optimal_k}
\alias{optimal_k}
\title{Optimal Number of Clusters based on Jensen-Shannon Divergence}
\usage{
optimal_k(X, kmin, kmax, seed_list = NULL)
}
\arguments{
\item{X}{A OTU matrix with dimensions \eqn{D \times p}.}

\item{kmin}{Numeric; the minimum number of clusters to be considered.}

\item{kmax}{Numeric; the maximum number of clusters to be considered.}

\item{seed_list}{List of numeric values; seeds for reproducibility for each k value, default is NULL.}
}
\value{
The optimal number of clusters, K, that maximizes the average Jensen-Shannon Divergence.
}
\description{
This function identifies the optimal number of clusters for fitting a zinck model
using Jensen-Shannon Divergence. The function fits the model for various values of clusters and calculates
the average Jensen-Shannon Divergence for each, returning the number of clusters that maximizes this value.
}
\examples{
\dontrun{
data("count.genus.RData")
dcount <- count.genus[,order(decreasing=T,colSums(count.genus,na.rm=T), apply(count.genus,2L,paste,collapse=''))]
X <- dcount
result <- optimal_k(X, kmin=8, kmax=11, seed_list=list(1,11,1,1))
print(result)
}
}
